todo-v1/requirements.txt
asgiref
certifi
charset-normalizer
dj-static
Django
django-cors-headers
djangorestframework
gunicorn
heroku
idna
python-dateutil
python-decouple
pytz
requests
six
sqlparse
static3
urllib3
